Move enrolments/activities PQA to Provider workspace
