+++ src/Server.UI/DependencyInjection.cs
@@ -15,8 +15,8 @@ using ActualLab.Fusion.Extensions;
 using Cfo.Cats.Server.UI.Middlewares;
 using Cfo.Cats.Server.UI.Hubs;
 using ApexCharts;
-using Cfo.Cats.Server.UI.Pages.QA.Enrolments;
-using Cfo.Cats.Server.UI.Pages.QA.Activities;
+using Cfo.Cats.Server.UI.Pages.Workspaces.Provider.Pages.Enrolments;
+using Cfo.Cats.Server.UI.Pages.Workspaces.Provider.Pages.Activities;
 using StackExchange.Redis;
 using Cfo.Cats.Application.Common.Interfaces.Identity;
 using Cfo.Cats.Infrastructure.Services.Identity;
