+++ src/Server.UI/Pages/Workspaces/Provider/Components/ProviderPageHeader.razor
@@ -30,6 +30,12 @@
             "case-management" => [home, ProviderLinks.CaseManagement.AsMudBreadcrumb()],
             "performance" => [home, ProviderLinks.Performance.AsMudBreadcrumb()],
             "cumulatives" => [home, ProviderLinks.Cumulatives.AsMudBreadcrumb()],
+            "enrolments/pqa" => [home, ProviderLinks.EnrolmentsPqa.AsMudBreadcrumb()],
+            "activities/pqa" => [home, ProviderLinks.ActivitiesPqa.AsMudBreadcrumb()],
+            _ when segment.StartsWith("enrolments/pqa/", StringComparison.Ordinal) =>
+                [home, ProviderLinks.EnrolmentsPqa.AsMudBreadcrumb(), new BreadcrumbItem("Enrolment", href: null, disabled: true)],
+            _ when segment.StartsWith("activities/pqa/", StringComparison.Ordinal) =>
+                [home, ProviderLinks.ActivitiesPqa.AsMudBreadcrumb(), new BreadcrumbItem("Activity", href: null, disabled: true)],
             "payments" => [home, ProviderLinks.Payments.AsMudBreadcrumb()],
             _ => throw new Exception()
         };
