+++ src/Server.UI/Pages/Workspaces/Provider/Pages/Enrolments/PQA.razor
@@ -1,4 +1,4 @@
-@page "/pages/qa/enrolments/pqa/{id:guid}"
+@page "/pages/workspace/provider/enrolments/pqa/{id:guid}"
 
 @attribute [Authorize(Policy = SecurityPolicies.Pqa)]
 @using Cfo.Cats.Application.Common.Validators
@@ -9,8 +9,11 @@
 
 @using Cfo.Cats.Application.SecurityConstants
 @using Cfo.Cats.Domain.Common.Enums
+@using Cfo.Cats.Server.UI.Pages.Workspaces.Provider.Components
 @using Cfo.Cats.Server.UI.Pages.Workspaces.ServiceDesk.Pages.Enrolments.Components
 @inherits CatsComponentBase
+
+<ProviderPageHeader />
 
 <style>
     .document-container {
